requirements.txt
tensorflow==2.12.0
datasets==2.12.0
numpy==1.24.4
Pillow==11.1.0